requirements.txt
numpy
pillow
sounddevice
soundfile
tk